Progress on slack bot and fixing tests
